requirements.txt
pymilvus
langchain
langchain-openai
langchain-core
langchain-community
langchain-text-splitters
openai
python-dotenv
psycopg2